Add VLCStyledVideoWindow.
This is deactivated by default. The goal is to allow to easily theme the window controller. The current default theme mimic an aqua interface. To create theme you have to use html/css/javascript. This is using WebKit as a backend.

You can turn this on, by activating that feature into VLCFeatures.h.