Fixes window zoom functionality. Closes gh-7.
This implementation fixes the functionality of the
zoom, however, because the window is skinned the 
window's real size is bigger than what is
displayed.  We may need  to create a callback to
the skin to ask the skin what should be the size
of the zoomed window or we determine what the 
offset to determine how large to the make the
window.