Save user settings on exit
Before if you exited before one of the random times the app flushes the data, you would loose your settings. Now it flushes on exit and if it fails, logs it. Please note that we may want to decide what we want it to do on failure. Should it create a pop-up error message? Should it just quit and log it?

In the mean time, this now uses VLCAssert